+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 
 * Fixed broken #find al (Alan Larkin)
+* Add `#find_by_slug!` al (Alan Larkin)
 
 ## 0.10.0
 * Add .find_unique_slug_for and #find_unique_slug_for methods DouweM (Douwe Maan)